# SwaggerV2.yaml
swagger: '2.0'
info:
  title: CLIMATE DATA REST API - OpenAPI 3.0
  description: |-
    This is an amazing REST-API developed by the UpTown IT team for allowing Students and Teachers to interact with Raw-Weather-Data collected from the Weather Station ATM41 (Realtime).
    The endpoints shown here have two purposes:
      1. Interact with Users (that could be administrators, teachers or students).
      2. Interact with Data obtained from the Weather Stations.
host: localhost:3000
basePath: /
paths:
...
  /data/readings/{id}:
    get:
      summary: /data/readings/:id (auth)
      tags:
      - Misc
      operationId: /data/readings/:id(auth-student)
      deprecated: false
      produces:
      - application/json; charset=utf-8
      parameters:
      - name: Authorization
        in: header
        required: true
        type: string
        description: ''
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/~1data~1readings~1%3Aid%28auth-student%29'
          examples:
            application/json; charset=utf-8:
              _id: 63f9aae72135908d2391bbb1
              Atmospheric Pressure (kPa): '100.00'
              Device Name: CUSTOM LOCATION FAKE DATA X
              Humidity (%): '50'
              Lightning Average Distance (km): '0.0'
              Lightning Strike Count: '0.0'
              Location: -55.00000, 55.00000
              Maximum Wind Speed (m/s): '4.00'
              Precipitation mm/h: '10.0'
              Solar Radiation (W/m2): '90.0'
              Temperature (°C): '40.0'
              Time: 2023-02-25T06:29:59.889Z
              Vapor Pressure (kPa): '2.00'
              Wind Direction (°): '300.0'
              Wind Speed (m/s): '2.00'
              __v: 0
          headers:
            X-Powered-By:
              type: string
              default: Express
            Access-Control-Allow-Origin:
              type: string
              default: '*'
            Content-Length:
              type: string
              default: '489'
            Date:
              type: string
              default: Sat, 25 Feb 2023 23:11:16 GMT
            Connection:
              type: string
              default: keep-alive
            Keep-Alive:
              type: string
              default: timeout=5
      security: []
    put:
      summary: /data/readings/:id 1 (auth-admin)
      tags:
      - Misc
      operationId: /data/readings/:id1(auth-admin)
      deprecated: false
      produces:
      - application/json; charset=utf-8
      parameters:
      - name: Authorization
        in: header
        required: true
        type: string
        description: ''
      - name: Body
        in: body
        required: true
        description: ''
        schema:
          $ref: '#/definitions/~1data~1readings~1%3Aid1%28auth-admin%29request'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/~1data~1readings~1%3Aid1%28auth-admin%29'
          examples:
            application/json; charset=utf-8:
              _id: 63f9aae72135908d2391bbb1
              Atmospheric Pressure (kPa): '100.00'
              Device Name: CUSTOM LOCATION FAKE DATA X
              Humidity (%): '50'
              Lightning Average Distance (km): '0.0'
              Lightning Strike Count: '0.0'
              Location: -55.00000, 55.00000
              Maximum Wind Speed (m/s): '4.00'
              Precipitation mm/h: '10.0'
              Solar Radiation (W/m2): '90.0'
              Temperature (°C): '40.0'
              Time: 2023-02-25T06:29:59.889Z
              Vapor Pressure (kPa): '2.00'
              Wind Direction (°): '300.0'
              Wind Speed (m/s): '2.00'
              __v: 0
          headers:
            X-Powered-By:
              type: string
              default: Express
            Access-Control-Allow-Origin:
              type: string
              default: '*'
            Content-Length:
              type: string
              default: '489'
            Date:
              type: string
              default: Sat, 25 Feb 2023 23:12:51 GMT
            Connection:
              type: string
              default: keep-alive
            Keep-Alive:
              type: string
              default: timeout=5
      security: []
    patch:
      summary: /data/readings/:id 1 (auth-admin)
      tags:
      - Misc
      operationId: /data/readings/:id1(auth-admin)
      deprecated: false
      produces:
      - application/json; charset=utf-8
      parameters:
      - name: Authorization
        in: header
        required: true
        type: string
        description: ''
      - name: Body
        in: body
        required: true
        description: ''
        schema:
          $ref: '#/definitions/~1data~1readings~1%3Aid1%28auth-admin%29request1'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/~1data~1readings~1%3Aid1%28auth-admin%29'
          examples:
            application/json; charset=utf-8:
              _id: 63f9aae72135908d2391bbb1
              Atmospheric Pressure (kPa): '100.00'
              Device Name: CUSTOM LOCATION FAKE DATA X
              Humidity (%): '50'
              Lightning Average Distance (km): '0.0'
              Lightning Strike Count: '0.0'
              Location: -55.00000, 55.00000
              Maximum Wind Speed (m/s): '4.00'
              Precipitation mm/h: '10.0'
              Solar Radiation (W/m2): '90.0'
              Temperature (°C): '40.0'
              Time: 2023-02-25T06:29:59.889Z
              Vapor Pressure (kPa): '2.00'
              Wind Direction (°): '300.0'
              Wind Speed (m/s): '2.00'
              __v: 0
          headers:
            X-Powered-By:
              type: string
              default: Express
            Access-Control-Allow-Origin:
              type: string
              default: '*'
            Content-Length:
              type: string
              default: '489'
            Date:
              type: string
              default: Sat, 25 Feb 2023 23:12:57 GMT
            Connection:
              type: string
              default: keep-alive
            Keep-Alive:
              type: string
              default: timeout=5
      security: []
    delete:
      summary: /data/readings/:id (auth-student)
      tags:
      - Misc
      operationId: /data/readings/:id(auth-student)
      deprecated: false
      produces:
      - application/json; charset=utf-8
      parameters:
      - name: Authorization
        in: header
        required: true
        type: string
        description: ''
      responses:
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/~1data~1readings~1%3Aid%28auth-student%292'
          examples:
            application/json; charset=utf-8:
              message: Unauthorized access
          headers:
            X-Powered-By:
              type: string
              default: Express
            Access-Control-Allow-Origin:
              type: string
              default: '*'
            Content-Length:
              type: string
              default: '33'
            Date:
              type: string
              default: Sat, 25 Feb 2023 23:13:11 GMT
            Connection:
              type: string
              default: keep-alive
            Keep-Alive:
              type: string
              default: timeout=5
      security: []
  /data/readings:
    post:
      summary: /data/readings (auth-admin)
      tags:
      - Misc
      operationId: /data/readings(auth-admin)
      deprecated: false
      produces:
      - application/json; charset=utf-8
      parameters:
      - name: Authorization
        in: header
        required: true
        type: string
        description: ''
      - name: Body
        in: body
        required: true
        description: ''
        schema:
          $ref: '#/definitions/~1data~1readings%28auth-student%29request'
      responses:
        '201':
          description: OK
          schema:
            $ref: '#/definitions/~1data~1readings%28auth-student%291'
          examples:
            application/json; charset=utf-8:
            - _id: 63f2fbb80be22c39ece6ee35
              Atmospheric Pressure (kPa): '100.00'
              Device Name: CUSTOM LOCATION FAKE DATA X
              Humidity (%): '70.0'
              Lightning Average Distance (km): '0.0'
              Lightning Strike Count: '0.0'
              Location: -30.00000, 30.00000
              Maximum Wind Speed (m/s): '4.00'
              Precipitation mm/h: '10.0'
              Solar Radiation (W/m2): '90.0'
              Temperature (°C): '40.0'
              Time: 2023-02-20T04:48:56.510Z
              Vapor Pressure (kPa): '2.00'
              Wind Direction (°): '300.0'
              Wind Speed (m/s): '2.00'
              __v: 0
              precipitation: 10
          headers:
            X-Powered-By:
              type: string
              default: Express
            Access-Control-Allow-Origin:
              type: string
              default: '*'
            Content-Length:
              type: string
              default: '33'
            Date:
              type: string
              default: Sat, 25 Feb 2023 23:12:33 GMT
            Connection:
              type: string
              default: keep-alive
            Keep-Alive:
              type: string
              default: timeout=5
      security: []
  /data/advanced/max-precipitation:
    get:
      summary: /data/advanced/max-precipitation (auth)
      tags:
      - Misc
      operationId: /data/advanced/max-precipitation(auth)
      deprecated: false
      produces:
      - application/json; charset=utf-8
      parameters:
      - name: Authorization
        in: header
        required: true
        type: string
        description: ''
      responses:
        '200':
          description: OK
          schema:
            type: array
            default: ''
            items:
              $ref: '#/definitions/~1data~1advanced~1max-precipitation%28auth%29'
          examples:
            application/json; charset=utf-8:
            - _id: 63f2fbb80be22c39ece6ee35
              Atmospheric Pressure (kPa): '100.00'
              Device Name: CUSTOM LOCATION FAKE DATA X
              Humidity (%): '70.0'
              Lightning Average Distance (km): '0.0'
              Lightning Strike Count: '0.0'
              Location: -30.00000, 30.00000
              Maximum Wind Speed (m/s): '4.00'
              Precipitation mm/h: '10.0'
              Solar Radiation (W/m2): '90.0'
              Temperature (°C): '40.0'
              Time: 2023-02-20T04:48:56.510Z
              Vapor Pressure (kPa): '2.00'
              Wind Direction (°): '300.0'
              Wind Speed (m/s): '2.00'
              __v: 0
              precipitation: 10
          headers:
            X-Powered-By:
              type: string
              default: Express
            Access-Control-Allow-Origin:
              type: string
              default: '*'
            Content-Length:
              type: string
              default: '512'
            Date:
              type: string
              default: Sat, 25 Feb 2023 23:13:32 GMT
            Connection:
              type: string
              default: keep-alive
            Keep-Alive:
              type: string
              default: timeout=5
      security: []
  /data/query:
    get:
      summary: /data/query?station&start&end (auth)
      tags:
      - Misc
      operationId: /data/query?station&start&end(auth)
      deprecated: false
      produces:
      - application/json; charset=utf-8
      parameters:
      - name: station
        in: query
        required: true
        type: string
        description: ''
      - name: start
        in: query
        required: true
        type: string
        description: ''
      - name: end
        in: query
        required: true
        type: string
        description: ''
      - name: Authorization
        in: header
        required: true
        type: string
        description: ''
      responses:
        '200':
          description: OK
          schema:
            type: array
            default: ''
            items:
              $ref: '#/definitions/~1data~1query%3Fstation%26start%26end%28auth%29'
          examples:
            application/json; charset=utf-8:
            - _id: 63f2fb320be22c39ece6ee2c
              Atmospheric Pressure (kPa): '100.00'
              Device Name: CUSTOM LOCATION FAKE DATA X
              Precipitation mm/h: '5.0'
              Solar Radiation (W/m2): '90.0'
              Temperature (°C): '40.0'
              Time: 2023-02-20T04:46:42.922Z
            - _id: 63f2fb360be22c39ece6ee2f
              Atmospheric Pressure (kPa): '100.00'
              Device Name: CUSTOM LOCATION FAKE DATA X
              Precipitation mm/h: '4.0'
              Solar Radiation (W/m2): '90.0'
              Temperature (°C): '40.0'
              Time: 2023-02-20T04:46:46.517Z
          headers:
            X-Powered-By:
              type: string
              default: Express
            Access-Control-Allow-Origin:
              type: string
              default: '*'
            Content-Length:
              type: string
              default: '475'
            Date:
              type: string
              default: Sat, 25 Feb 2023 23:13:42 GMT
            Connection:
              type: string
              default: keep-alive
            Keep-Alive:
              type: string
              default: timeout=5
      security: []
...